add label if device deleted. add events than was deleted because device is deleted

+++ resources/views/devices/index.blade.php
@@ -10,12 +10,15 @@
                             Device ID: {{ $device->id }} <br>
                             Device name: {{ $device->name }} <br>
                             <div class="pull-right">
-
-                                {{ Form::open(['route' => ['devices.destroy', $device->id], 'method' => 'delete', 'style' => 'text-align:right']) }}
-                                <a class="btn btn-primary" href="{{route('devices.edit', ['id' => $device->id])}}" class="alert-link">Edit</a>
-                                {{ Form::submit('Delete', ['class'=>'btn btn-danger']) }}
-                                {{ Form::close() }}
-
+                                @if (!isset($device->deleted_at))
+                                    {{ Form::open(['route' => ['devices.destroy', $device->id], 'method' => 'delete', 'style' => 'text-align:right']) }}
+                                    <a class="btn btn-primary" href="{{route('devices.edit', ['id' => $device->id])}}"
+                                       class="alert-link">Edit</a>
+                                    {{ Form::submit('Delete', ['class'=>'btn btn-danger']) }}
+                                    {{ Form::close() }}
+                                @else
+                                    DELETED
+                                @endif
                             </div>
                         </div>
                     </div>
